rewind: Revert: firebase 코드 복구
